RPM Control added (with SpringAnimation)
Now has manual control over RPM value (to simulate CAN message etc), using the SpringAnimation add on.

The animation is jerky and not quite right on Mac OS but for some reason runs well on Windows. Seems to be a Qt bug that hopefully won't carry over to Embedded Linux on the Pi.