Fixed Bugs and Made some Button Connections
Fixed Bug where Window wouldn't open. Added a few more Icons, Including a Home Button. Setup a state switch so that the BMW image is not shown on the press of settings button, and reappears on Home. Added car unlock and lock button sequence so that the Door button triggers the Lock and unlock images respectively.