Lots of debugging and tweaks, Simulink model now executes
The background Simulink model now executes. It currently doubles a value that has been sent to it via the QML, but does so on a 100Hz cycle, as intended.

Now left is to connect the input and output from the slider in the QML, through to Simulink, and then back over to the dial to prove that the connection can be made, then the mass spring damper model can be implemented

Also, a fresh build from Simulink incurs one #include error which is due to the nested nature of that particular file. A method of instructing the Qt project of this file's location outside of the header file would mean that the Simulink code doesn't have to be modified after each build.